refactor(rvr): move ext_hint_stream_set function from RV64I extension to core rvr
